% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageMSE.R
\docType{data}
\name{averageMSE}
\alias{averageMSE}
\title{averageMSE}
\format{A \code{data.frame}}
\usage{
averageMSE
}
\description{
averageMSE
}
\details{
The object is used internally, from mahalData
}
\keyword{datasets}
