\name{csci-package}
\alias{csci-package}
\alias{csci-package}
\docType{package}
\title{
\packageTitle{csci}
}
\description{
\packageDescription{csci}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{csci}
\packageIndices{csci}
The package only has one main function \code{\link{CSCI}} and one data set \code{\link{hepABulg}}.
}
\author{
\packageAuthor{csci}

Maintainer: \packageMaintainer{csci}
}
%\references{}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{~~ simple examples of the most important functions ~~}
