% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\docType{data}
\name{extent}
\alias{extent}
\title{Sample extents}
\format{A \code{data.frame}}
\usage{
extent
}
\description{
Sample extents
}
\details{
The object is used internally
}
\examples{
data(extent)
}
\keyword{datasets}
