% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_forests.R
\docType{data}
\name{final_forests}
\alias{final_forests}
\title{Random forest models}
\format{A list of \code{\link[randomForest]{randomForest}} modela}
\usage{
final_forests
}
\description{
Random forest models used for prediction of CSCI O/E scores
}
\examples{
data(final_forests)
}
\keyword{datasets}
