% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetaData.R
\name{loadMetaData}
\alias{loadMetaData}
\title{Read in reference taxonomy data used to build CSCI models}
\usage{
loadMetaData(metadata.year = 2025)
}
\description{
Read in reference taxonomy data used to build CSCI models
}
\details{
This function loads a data frame containing all taxonomic data at reference sites used to calibrate and validate the CSCI, including their sample date, 
FinalID, BAResult, LifeStageCode, SiteSet (RefCal, RefVal), and SAFIT designations.
}
