% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRefBugData.R
\name{loadRefBugData}
\alias{loadRefBugData}
\title{Read in reference taxonomy data used to build CSCI models}
\usage{
loadRefBugData()
}
\description{
Read in reference taxonomy data used to build CSCI models
}
\details{
This function loads a data frame containing all taxonomic data at reference sites used to calibrate and validate the CSCI, including their sample date, 
FinalID, BAResult, LifeStageCode, SiteSet (RefCal, RefVal), and SAFIT designations.
}
