% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingMidges.R
\name{MissingMidges}
\alias{MissingMidges}
\title{CSCI estimator for samples taken to SAFIT Level 1}
\usage{
MissingMidges(mylist)
}
\arguments{
\item{mylist}{An output from the CSCI object (i.e., a list of dataframes)}
}
\value{
Returns the same dataframe but with new fields added to the core report, with "_MisingMidges" appended to the field name.
}
\description{
CSCI estimator for samples taken to SAFIT Level 1
}
\details{
This function estimates maximum O/E and CSCI scores if all "missing" midges are present, i.e., SAFIT Leve 1.
}
