% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refprox.R
\name{refprox}
\alias{refprox}
\title{Plot reference proximiy values for a test site}
\usage{
refprox(station_dat, output = c("map", "jit", "pca", "dat"))
}
\arguments{
\item{station_dat}{input \code{data.frame} for the test site that includes all station data}

\item{output}{chr string indicating desired output from the function, one of \code{"map", "jit", "pca", "dat"}}
}
\value{
A plot of average proximity values across all metrics if \code{output = "map"}, a facetted map for each 
metric of proximity values if \code{output = "fac"}, a PCA plot of principal components one and two for 
all metric proximity values if \code{output = "pca"}, or a \code{data.frame} of the proximity values for 
all reference sites and the test site if \code{output = "dat"}.
}
\description{
Plot reference proximiy values for a test site
}
\details{
\code{station_dat} must have the same GIS predictors as those in \code{bugs_stations[[2]]}, specifically 
\code{"StationCode", "AREA_SQKM", "New_Lat", "New_Long", "SITE_ELEV", "PPT_00_09", "TEMP_00_09", "SumAve_P", 
"KFCT_AVE", "BDH_AVE", "P_MEAN", "ELEV_RANGE"}. The \code{StationCode} must also not be a reference site, as 
in those in \code{loadRefData()}.
}
\examples{
station_dat <- bugs_stations[[2]][1, ]

# map, default
refprox(station_dat)

# jitter
refprox(station_dat, output = 'jit')

# pca
refprox(station_dat, output = 'pca')

# data
refprox(station_dat, output = 'dat')
}
