% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refsamples.R
\docType{data}
\name{refsamples}
\alias{refsamples}
\title{refsamples}
\format{A \code{data.frame}}
\usage{
refsamples
}
\description{
refsamples
}
\details{
The object is used internally
}
\keyword{datasets}
