% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supp1_OE.R
\name{Supp1_OE}
\alias{Supp1_OE}
\title{A cheater function for getting only the Supp1_OE scores}
\usage{
Supp1_OE(bugs, stations, rand = sample.int(10000, 1))
}
\arguments{
\item{bugs}{input taxonomy data}

\item{stations}{input station data}

\item{rand}{random integer for subsampling}
}
\value{
A data.frame for the Supp1_OE report
}
\description{
A cheater function for getting only the Supp1_OE scores
}
\examples{
data(bugs_stations) 
Supp1_OE(bugs = bugs_stations[[1]], stations = bugs_stations[[2]])
}
