% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRefData.r
\name{loadRefData}
\alias{loadRefData}
\title{Read in reference data used to build CSCI models}
\usage{
loadRefData()
}
\description{
Read in reference data used to build CSCI models
}
\details{
This function loads a data frame containing all reference sites used to calibrate and validate the CSCI, including their sample date, 
biotic group ID, O, E, O/E, MMI scores, observed metrics, predicted metrics, scored metrics, and PSA regions.
}
